for consistency, put PhotoServiceApiClientImpl into own file
